Rename store to article-store.
