+++ tests/scenarios/displaying-article-overview.test.js
@@ -21,7 +21,7 @@ describe('displaying article overview', () => {
 
     var ActionCreator = require('../../app/action-creator.js');
     var actionCreator = new ActionCreator(dataAccess);
-    var Store = require('../../app/store.js');
+    var Store = require('../../app/article-store.js');
     var store = new Store(actionCreator);
     var renderedComponent = React.render(<ComponentClass actionCreator={actionCreator} store={store}/>, renderTarget);
     component = renderedComponent;
